pyproject.toml
[project]
name = "auth-app"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi>=0.121.1",
    "httpx>=0.28.1",
    "passlib>=1.7.4",
    "pydantic>=2.12.4",
    "pyjwt>=2.10.1",
    "pytest>=9.0.0",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
